requirements.txt
transformers
torch
faiss-cpu
sentence-transformers
torch-geometric
networkx
